chore: slightly more optimized
