+++ src/services/variableHelpers.ts
@@ -1,23 +1,23 @@
-import { AdHocFiltersVariable, sceneGraph, SceneObject } from '@grafana/scenes';
+import { AdHocFiltersVariable, AdHocFilterWithLabels, sceneGraph, SceneObject } from '@grafana/scenes';
 
 import { IndexScene } from '../Components/IndexScene/IndexScene';
 import { ServiceScene } from '../Components/ServiceScene/ServiceScene';
 import { CustomConstantVariable } from './CustomConstantVariable';
 import { FilterOp } from './filterTypes';
-import { LabelFiltersVariable } from './LabelFiltersVariable';
 import { isOperatorInclusive } from './operatorHelpers';
 import { includeOperators, numericOperators, operators } from './operators';
+import { ReadOnlyAdHocFiltersVariable } from './ReadOnlyAdHocFiltersVariable';
 import { getRouteParams } from './routing';
 import { getLabelsVariable } from './variableGetters';
 import { SERVICE_NAME, SERVICE_UI_LABEL, VAR_LABELS } from './variables';
 
-type ClearableVariable = AdHocFiltersVariable | CustomConstantVariable | LabelFiltersVariable;
+type ClearableVariable = AdHocFiltersVariable | CustomConstantVariable | ReadOnlyAdHocFiltersVariable;
 export function getVariablesThatCanBeCleared(indexScene: IndexScene): ClearableVariable[] {
   const variables = sceneGraph.getVariables(indexScene);
   let variablesToClear: ClearableVariable[] = [];
 
   for (const variable of variables.state.variables) {
-    if (variable.state.name === VAR_LABELS && variable instanceof LabelFiltersVariable) {
+    if (variable instanceof ReadOnlyAdHocFiltersVariable) {
       if (
         !variable.state.filters.every((filter) =>
           variable
@@ -52,24 +52,19 @@ export function clearVariables(sceneRef: SceneObject) {
   const variablesToClear = getVariablesThatCanBeCleared(indexScene);
 
   variablesToClear.forEach((variable) => {
-    if (variable instanceof LabelFiltersVariable) {
+    if (variable instanceof ReadOnlyAdHocFiltersVariable) {
       let { labelName } = getRouteParams(sceneRef);
       // labelName is the label that exists in the URL, which is "service" not "service_name"
       if (labelName === SERVICE_UI_LABEL) {
         labelName = SERVICE_NAME;
       }
       const readonlyFilters = variable.getReadonlyFilters();
+      const readonlyFiltersSet = new Set<string>();
+      addFiltersToSet(readonlyFilters ?? [], readonlyFiltersSet);
       variable.setState({
-        filters: variable.state.filters.filter(
-          (filter) =>
-            filter.key === labelName ||
-            readonlyFilters?.find(
-              (readonlyFilter) =>
-                readonlyFilter.key === filter.key &&
-                readonlyFilter.value === filter.value &&
-                readonlyFilter.operator === filter.operator
-            )
-        ),
+        filters: variable.state.filters.filter((filter) => {
+          return filter.key === labelName || readonlyFiltersSet.has(getFilterSetKey(filter));
+        }),
       });
     } else if (variable instanceof AdHocFiltersVariable) {
       variable.setState({
@@ -114,6 +109,11 @@ export const operatorFunction = function (variable: AdHocFiltersVariable) {
 
   return operators;
 };
+
+export const getFilterSetKey = (filter: AdHocFilterWithLabels) =>
+  filter.key + '_' + filter.operator + '_' + filter.value;
+export const addFiltersToSet = (filters: AdHocFilterWithLabels[], set: Set<string>) =>
+  filters.forEach((filter) => set.add(getFilterSetKey(filter)));
 
 /**
  * For embedded contexts, return the first label as primary label
